// src/app/interfaces/food-entry.ts
export interface FoodEntry {
  foodId?: number;
  servingSize: number;
  foodName?: string;
  calories?: number;
  protein?: number;
  fat?: number;
  carbs?: number;
  cholesterol?: number;
  sodium?: number;
  fiber?: number;
  sugar?: number;
  addedSugar?: number;
  vitaminD?: number;
  calcium?: number;
  iron?: number;
  potassium?: number;
  notes?: string;
}